requirements.txt
appdirs==1.4.4
boto3==1.24.28
botocore==1.27.59
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.6
cryptography==39.0.1
Flask==2.2.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.0.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numexpr==2.8.4
numpy==1.22.3
packaging==23.0
pandas==1.5.3
pip==23.1.2
pooch==1.4.0
pycparser==2.21
pymongo==3.12.0
pyOpenSSL==23.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7
requests==2.29.0
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
setuptools==67.8.0
six==1.16.0
threadpoolctl==2.2.0
urllib3==1.26.16
Werkzeug==2.2.3
wheel==0.38.4
win-inet-pton==1.1.0
zipp==3.11.0
